requirements.txt
flask==2.3.3
flask-wtf==1.1.1
wtforms==3.0.1
supabase==1.0.4
python-dotenv==1.0.0
cryptography==41.0.7
werkzeug==2.3.7
email_validator==2.3.0